chore(RootContainer): moves App code to RootContainer
