--- AppContainer.js
@@ -1,18 +1,0 @@
-// import { createSwitchNavigator, createAppContainer } from 'react-navigation';
-// import AuthNavigation from './AuthNavigation';
-// import Navigation from './AppNavigation';
-// import StackContainer from './StackContainer';
-
-// const SwitchNavigator = createSwitchNavigator(
-//   {
-//     Auth: AuthNavigation,
-//     App: Navigation,
-//   },
-//   {
-//     initialRouteName: 'Auth',
-//   }
-// );
-
-// const AppContainer = createAppContainer(SwitchNavigator);
-
-// export default AppContainer;
